fix(analytics): changes-made-for-new-design
