+++ src/components/analytics/Inventory.vue
@@ -1,4 +1,22 @@
 <template>
+  <div class="flex-box">
+    <label class="pad-label ls" for="start_date">
+      <strong>Company:</strong>
+    </label>
+    <select
+      id="company-type"
+      name="company-type"
+      class="custom-select"
+      style="width: 30%"
+      v-model="company"
+      @change="fetchAnalyticsBtn"
+      :disabled="!admin"
+    >
+      <option class="batches-op" v-for="company in companies" v-bind:key="company.id" v-bind:value="company.id">
+        {{company.company_name}}
+      </option>
+    </select>
+  </div>
   <table>
     <colgroup>
       <col span="1" style="width: 50%;">
@@ -17,14 +35,57 @@
 
 <script lang="ts">
 import {defineComponent} from 'vue';
-import { mapGetters } from 'vuex';
+import { mapGetters, mapActions } from 'vuex';
+import { ActionTypes } from '@/store/modules/order/actions';
+import { ActionTypes as AuthActionTypes } from '@/store/modules/auth/actions';
 
 export default defineComponent({
   name: 'InventoryAnaltyics',
+  data() {
+    const date = new Date();
+    const dateStr = `${date.getFullYear()}-${date.getMonth()+1}-${date.getDate()}`;
+    return {
+      startDate: dateStr,
+      endDate: dateStr,
+      company: 0,
+    };
+  },
   computed: {
     ...mapGetters({
       analytics: 'getAnalytics',
-    })
+      companies: 'getInventoryCompanies',
+      userdata: 'getUser',
+    }),
+    admin(){
+      const allowedRoles = ['SUPER_ADMIN', 'ADMIN'];
+      const allowedCompanies = ['PARENT', 'STORE'];
+      if(this.userdata != null && allowedRoles.includes(this.userdata.user_type)
+        && allowedCompanies.includes(this.userdata.company.company_type) 
+      ){
+        return true;
+      }
+      return false;
+    },
+  },
+  methods: {
+    ...mapActions({
+      fetchAnalytics: ActionTypes.FETCH_ANALYTICS,
+      fetchCompanies: AuthActionTypes.FETCH_ALL_COMPANIES,
+      fetchUser: AuthActionTypes.USER_DATA,
+    }),
+    async fetchAnalyticsBtn() {
+      await this.fetchAnalytics({
+        start_date: this.startDate,
+        end_date: this.endDate,
+        company: this.company,
+      });
+    }
+  },
+  async mounted() {
+    await this.fetchUser();
+    await this.fetchCompanies();
+    await this.fetchAnalyticsBtn();
+    this.company = this.userdata.company.id;
   }
 })
 </script>
