+++ src/router.ts
@@ -253,11 +253,6 @@ const routes: Array<RouteRecordRaw> = [
         component: OrderAnaltyics,
       },
       {
-        path: 'finance',
-        name: 'finance-analytics',
-        component: FinanceAnaltyics,
-      },
-      {
         path: 'inventory',
         name: 'inventory-analytics',
         component: InventoryAnaltyics,
