+++ src/store/modules/order/actions.ts
@@ -91,7 +91,7 @@ export interface Actions {
   [ActionTypes.UPDATE_ORDER]({ commit }: AugmentedActionContext, order: Order): void;
   [ActionTypes.SET_FIELD_ERROR]({ commit }: AugmentedActionContext, error: any): void;
   [ActionTypes.EMPTY_ORDER]({ commit }: AugmentedActionContext, error: any): void;
-  [ActionTypes.FETCH_ANALYTICS]({ commit }: AugmentedActionContext, options: { start_end: Date; end_date: Date}): void;
+  [ActionTypes.FETCH_ANALYTICS]({ commit }: AugmentedActionContext, options: { start_end: Date; end_date: Date; company: number }): void;
 }
 
 export const actions: ActionTree<State, IRootState> &
@@ -380,7 +380,7 @@ Actions = {
   async [ActionTypes.EMPTY_ORDER]({ commit }: AugmentedActionContext, error: any) {
     commit(MutationTypes.SetOrder, {});
   },
-  async [ActionTypes.FETCH_ANALYTICS]({ commit }: AugmentedActionContext, options: {start_end: Date; end_date: Date}) {
+  async [ActionTypes.FETCH_ANALYTICS]({ commit }: AugmentedActionContext, options: {start_end: Date; end_date: Date; company: number }) {
     const response = await serverRequest('get', 'analytics/', true, undefined, options);
     if (isAxiosResponse(response)) {
       commit(MutationTypes.SetAnalytics, response.data);
